+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 =========
 
 * :py:func:`~core.uniquelevel` or :py:code:`.idx.unique` returns the unique values of one
-  or multiple levels. :pull:`7`
+  or multiple levels. :pull:`8`
 
 v0.2.2 (2023-05-02)
 ------------------------------------------------------------