+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 * :py:func:`~core.uniquelevel` or :py:code:`.idx.unique` returns the unique values of one
   or multiple levels. :pull:`9`
 * :py:func:`~core.summarylevel` creates a string summarizing the index levels and their
-  values. Can also be accessed as :py:code:`df.idx` or :py:code:`index.idx.` :pull:`8`
+  values. Can also be accessed as :py:code:`df.idx` or :py:code:`index.idx.` :pull:`9`
 
 v0.2.2 (2023-05-02)
 ------------------------------------------------------------